CAL-15: Create a lib package and sanitize the components
