+++ app/package.json
@@ -20,6 +20,7 @@
     "dagre-d3": "^0.6.3",
     "html-entities": "^1.3.1",
     "isarray": "^2.0.5",
+    "lib": "file:src/lib",
     "material-design-icons-iconfont": "^5.0.1",
     "robust-point-in-polygon": "^1.0.3",
     "socket.io-client": "^2.3.1",
