+++ app/src/components/callsiteCorrespondence/markers.vue
@@ -10,7 +10,7 @@
 
 <script>
 import * as d3 from "d3";
-import * as utils from "../utils";
+import * as utils from "lib/utils";
 
 export default {
 	name: "Markers",
