+++ app/src/components/ensembleHistogram/ensembleHistogram.vue
@@ -22,11 +22,16 @@
 </template>
 
 <script>
+// Library imports
 import * as d3 from "d3";
 import "d3-selection-multi";
+
+// Local library imports
+import * as utils from "lib/utils";
+import EventHandler from "lib/routing/EventHandler";
+
+// Local component imports
 import ToolTip from "./tooltip";
-import * as utils from "../utils";
-import EventHandler from "../EventHandler";
 
 export default {
 	name: "EnsembleHistogram",
