+++ app/src/components/sankey/encodings/meanDiff.vue
@@ -11,8 +11,6 @@
 
 <script>
 import * as d3 from "d3";
-import * as utils from "../../utils";
-import { color } from "dagre-d3/dist/dagre-d3";
 
 export default {
 	name: "MeanDiff",
