+++ app/src/components/singleHistogram/singleHistogram.vue
@@ -18,13 +18,17 @@
 </template>
 
 <script>
+// Library imports
 import * as d3 from "d3";
 import "d3-selection-multi"; 
-
-import ToolTip from "./tooltip";
-import * as utils from "../utils";
-import EventHandler from "../EventHandler";
 import { brush } from "d3";
+
+// Local library imports
+import * as utils from "lib/utils";
+import EventHandler from "lib/routing/EventHandler";
+
+// Local component imports
+import ToolTip from "./tooltip";
 
 export default {
 	name: "SingleHistogram",
